OSIS-500 Move _append_entity_version function to business file
